requirements.txt
Genshi==0.7.5
greenlet==1.1.0
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
six==1.16.0
SQLAlchemy==1.4.21
python-dotenv==0.18.0